Added GLFW library for implmeneting Window System.
Window System:
Has one Window.h interface abstraction, which should be implemented in the appropriate Window class for our platform.
As we will try to support all platfroms, we should abstract from all restrctions that could be.

WindowsWindow:
Added to platform/windows packages as Window class for Windows platform.
This implmentation now handles creating, clearing and destroying the window.
Also it is using the events from events package to link this with Event System of GLFW library.
EventDispatcher will be called in this class to dispatch event for onEvent and onWindowsClose functions.
Also, this class is handling errors from GLFW library, that can comes, while init or callbacking processes.

Events that will be tracked now into the console:
all MouseEvent, all KeyEvent.