Layer System: Now everything is splitted into Layers as , we should manage, what kind of update is better to render.
Also we have a LayerStack that is wrapper under vector structure and has it's own iterator to iterate over this vector. Has ability to push and pop layers and overlays. Overlay is also a layer but should be psuhed always in the end of our LayerStack.
When LayerStack destroys, it will also clear it's vector of layers. However, we have ability to pop the layer from this stack and we will not destroy it, so we can change it and push it later.

Hazel Application:
Now has methods to psuh overlays and layers to the stack. Also it has reference to the LayerStack, so in future we can manipulate with it in different ways.

Hazel module:
Now provides a namespace for Layer, which can be  implemented on the client side.

Sanbox Application:
Now has a simple example of Layer class called ExampleLayer.

ExampleLayer:
Just logs into a console onEvent and onUpdate methods.